Integrate tuples into type checker (#201)
* Work towards sized collections (tuples)

* Create separate Tuple Expect

- Take into account tuple sizes
- For tuples, in type checker, take into account
  individual tuple field types
- Handle tuples of identifiers

* Add recursive substitution for tuples

- Add log message for generating unions within the
  unification stage.
- If access on tuple, create union for access and
  every item of tuple.

Still getting certain errors, particularly in the
for control flow test.

* Add permutate to lock file

* Fix some mistakes in constr_col

Collections are now properly constrained.
The substitution seems to struggle with
identifiers, however.

Some other cleanup:
- Use usize in CaretPos
- Improve logging of substitutions
- Some other minor naming things

* Use environment to detect shadowed variables

We substitute variables which shadow a previous
variable with a variable which the user cannot
create (because it contains an illegal character).

We also store this substitution such that any
other uses of this variable which shadowed the
prior one get the new name.

This is far more elegant than the ad-hoc system
of storing variables, and removes some of the
complexity away from the constraints.

* Finish writing mapping function

- Rename _as to aliases in Node
- Remove generation in unification stage.
  Ideally, we should never have to generate
  anything here, all should be done in the
  generation stage as we can determine everything
  statically anyway.

* Fix constraint order gen in lookup

* Fix underflow errors for tokens

Remove iterating over a list of tuples
test for now to verify the build is still
functioning.

- Show msg in gen
- Re-add logic for expected.
  Specifically, primitives disguised
  as expressions.
  This should be caught during the
  generation stage, but I'm not sure
  why it isn't.
  Need to investigate.

* [ci skip] Extra tuple substitution rules

In the unification stage that is.
Not sure if these are strigly necessary,
or if this is a hack which will resolve
itself once we hopefully land on an
elegant solution.

It does appear to fix some substitution
issues.
However, we again run into the issues that
trivial constraints (such as primitive
expressions) are no longer generated in the
unification stage.

Ideally, all constraints should be
generated in the generation stage, and
the unification stage should solve these
as-is.

* Fix unification of tuples

The fix was: first substitute,
then proceed to extract the
tuple elements as necessary.
I think this is the only type of
expression (for now) where we need
to extract the individual elements.
All other expressions can be
dealt with as-is, and should be
simple enough that all constraints
are generated during the generation
stage.

We also ignore two tuples,
and just let the substitution do
its work in this case.

- Update some error msg formatting
- Simplify unify expression rules

* Add var mapping logic to union and intersect

Also move environment to generation stage, as it
is only used here.
It does not need to be exposed to the outside
world.

Expect(ed) now takes a mapping of variables, which
is a mapping of a string to a string, as it is all
that it needs for shadowing purposes.
We do not need to expose the entire environment to
Expect(ed).

+++ src/lex/token.rs
@@ -13,9 +13,9 @@ impl Lex {
     pub fn new(pos: &CaretPos, token: Token) -> Self {
         let start = pos.clone();
         let end = if let Token::Str(_str, _) = &token {
-            pos.clone().offset_line(max(_str.lines().count() as i32 - 1, 0))
+            pos.clone().offset_line(max((_str.lines().count() as i32 - 1) as usize, 0))
         } else if let Token::DocStr(_str) = &token {
-            pos.clone().offset_line(max(_str.lines().count() as i32 - 1, 0))
+            pos.clone().offset_line(max((_str.lines().count() as i32 - 1) as usize, 0))
         } else {
             pos.clone()
         };
@@ -134,8 +134,8 @@ pub enum Token {
 }
 
 impl Token {
-    pub fn width(self) -> i32 {
-        (match self {
+    pub fn width(self) -> usize {
+        match self {
             Token::Id(id) => id.len(),
             Token::Real(real) => real.len(),
             Token::Int(int) => int.len(),
@@ -145,7 +145,7 @@ impl Token {
             Token::DocStr(_str) => _str.len() + 6,
             Token::ENum(num, exp) => num.len() + 1 + exp.len(),
             other => format!("{}", other).len()
-        } as i32)
+        }
     }
 
     pub fn same_type(left: &Token, right: &Token) -> bool {
